Refactor locationDetails images to be its own component
